pyproject.toml
[tool.poetry]
name = "event-system"
version = "0.1.0"
description = "Event propagation and consumption system"
authors = ["Miglė Šimavičiūtė <simaviciut@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
pyyaml = "^6.0"
requests = "^2.31.0"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
sqlalchemy = "^2.0.25"
pydantic = "^2.6.0"

[tool.poetry.dev-dependencies]
pytest = "^7.4.4"
black = "^24.1.0"
isort = "^5.13.2"